+++ 2023winterbootcamp/src/App.tsx
@@ -1,14 +1,15 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { RecoilRoot } from "recoil";
-import Layout from "./components/Layout";
-import Main from "./Mainpage";
-import Mypage from "./Mypage";
-import Choose from "./Choosepage";
-import Interview from "./Interviewpage";
-import Footer from "./components/Footer";
-import Resultpage from "./Resultpage";
-import Endpage from "./Endpage";
+import React from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import { RecoilRoot } from 'recoil';
+import Layout from './components/Layout';
+import Main from './Mainpage';
+import Mypage from './Mypage';
+import Choose from './Choosepage';
+import Interview from './Interviewpage';
+import Footer from './components/Footer';
+import Resultpage from './Resultpage';
+import Endpage from './Endpage';
+import Startpage from './Startpage';
 
 function App() {
   return (
@@ -16,12 +17,16 @@ function App() {
       <Router>
         <Layout>
           <Routes>
-            <Route path="/" element={<Main />} />
-            <Route path="/mypage" element={<Mypage />} />
-            <Route path="/choose" element={<Choose />} />
-            <Route path="/interview/:id" element={<Interview />} />
-            <Route path="/end/:id" element={<Endpage interview_name="삼성전자 2022 하반기 3급"/>} />
-            <Route path="/result/:id" element={<Resultpage />} />
+            <Route path='/' element={<Main />} />
+            <Route path='/mypage' element={<Mypage />} />
+            <Route path='/choose' element={<Choose />} />
+            <Route path='/start/:id' element={<Startpage />} />
+            <Route path='/interview/:id' element={<Interview />} />
+            <Route
+              path='/end/:id'
+              element={<Endpage interview_name='삼성전자 2022 하반기 3급' />}
+            />
+            <Route path='/result/:id' element={<Resultpage />} />
           </Routes>
           <Footer />
         </Layout>
